Remove unnecessary `consumes` annotation element from Web API
`consumes = APPLICATION_JSON_VALUE` unnecessarily narrows down potential callers to only those that provide `application/json` as Content-Type. Others, like e.g. `text/plain` that is used be the Web Console are rejected. As those methods do not make use of the request body at all this limitation can not only be removed but enables the usage of the Web Console.