Make time to live for cache entries configurable
I'm using InMemoryCacheManager because it is a simple implementation of
a FF4JCacheManager which nevertheless satifies my requirements. The only
concession that it makes is the long time to live. I would like to have
a shorter time to live. This is now possible by creating the cache
manager with an arbitrary time to live.