Fix issue with case insensitive databases
When using PropertyStoreSpringJdbc together with DBs that
are case insensitive, like PostgreSQL, an exception was
being thrown when trying to create the schema, if it had
already been created before.

This commit leverages the existing JdbcUtils that
addresses properly the issue.

Also this commit updates PropertyStoreMongo to use
`getCollection` instead of `createCollection` when
creating schema because Fakemongo does not implement
a `createCollection` method, leads to exceptions.
`getCollection` will create a collection if didn't exist
 as per the documentation https://docs.mongodb.com/manual/reference/method/db.getCollection/#behavior

 Closes https://github.com/ff4j/ff4j/issues/446